+++ src/components/app/app.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import {App} from '../app/app.jsx';
+import App from '../app/app.jsx';
 
 it(`App correctly renders after relaunch`, () => {
