+++ src/index.js
@@ -1,11 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import {App} from './components/app/app.jsx';
+import offers from './mocks'
+import App from './components/app/app.jsx';
 
 const init = () => {
   ReactDOM.render(
-      <App/>,
+      <App
+      offers={offers}/>,
       document.getElementById(`root`)
   );
 };
